requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
apturl==0.5.2
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
asyncer==0.0.7
attrs==24.3.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.4
Brlapi==0.8.3
build==1.2.2.post1
cachetools==5.5.0
cbor==1.0.0
certifi==2020.6.20
chainlit==2.0.4
chardet==4.0.0
charset-normalizer==3.4.1
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.6.7
datasets==2.19.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.15
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
durationpy==0.9
evdev==1.6.1
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
fastapi==0.115.6
fasteners==0.14.1
filelock==3.16.1
filetype==1.2.0
FlagEmbedding==1.3.3
flatbuffers==24.12.23
frozenlist==1.5.0
fsspec==2024.3.1
future==0.18.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
idna==3.3
ijson==3.3.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
inscriptis==2.5.3
ir_datasets==0.5.9
jeepney==0.7.1
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.8.0
keyring==23.5.0
kubernetes==31.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-text-splitters==0.3.5
langsmith==0.2.11
language-selector==0.1
launchpadlib==1.10.16
Lazify==0.4.0
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
literalai==0.1.103
lockfile==0.12.2
louis==3.20.0
lxml==5.3.0
lz4==4.3.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
mdurl==0.1.2
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.25.2
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.20.1
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.15
overrides==7.7.0
packaging==23.2
pandas==2.2.3
paramiko==2.9.3
peft==0.14.0
pexpect==4.8.0
Pillow==9.0.1
posthog==3.8.3
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pyarrow==19.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyGObject==3.42.1
PyJWT==2.10.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pynput==1.7.8
pyparsing==2.4.7
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
python-xlib==0.33
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==3.6.8
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
sentence-transformers==3.4.0
sentencepiece==0.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.41.3
streamlit==1.42.0
sympy==1.13.1
syncer==2.0.3
systemd-python==234
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.2.1
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.44.2
trec-car-tools==2.6
triton==3.1.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unlzw3==0.2.3
uptrace==1.28.2
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.25.0
uvloop==0.21.0
wadllib==1.3.6
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchdog==6.0.0
watchfiles==0.20.0
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
xdg==5
xkit==0.0.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zlib-state==0.1.9
